requirements.txt
numpy>=1.24.0
xarray>=2023.01.0
scipy>=1.10.0
numba>=0.57.0
matplotlib>=3.7.0
pandas>=2.0.0
pyyaml>=6.0

rasterio>=1.3.0
geopandas>=0.12.0
cupy-cuda12x>=12.0.0

